pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hydropilot"
version = "0.1.0"
description = "A hydrological model driver framework for coupling models (SWAT, APEX, HBV, VIC) with UQPyL"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
hydropilot-validate = "hydro_pilot.cli.validate:main"

[project.optional-dependencies]
uqpyl = ["UQPyL"]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["hydro_pilot*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
